requirements.txt
yfinance==0.2.36
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1
newsapi-python==0.2.7 